@@ -6,7 +6,7 @@ repository: https://github.com/knyghtryda/flutter_anywhere_menus
 issue_tracker: https://github.com/knyghtryda/flutter_anywhere_menus/issues
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=0.1.4 <2.0.0"
 
 dependencies:
@@ -15,7 +15,7 @@ dependencies:
     sdk: flutter
 
 dev_dependencies:
-  pedantic: ^1.9.0
+  pedantic: ^1.11.1
   flutter_test:
     sdk: flutter
   
