@@ -9,10 +9,10 @@ part 'offset_methods.dart';
 
 /// This is the what contains both the widget you want to draw and the [MenuBar]
 class Menu extends StatefulWidget {
-  final Widget child;
+  final Widget? child;
 
   /// This contains all of the info for rendering the Menu
-  final MenuBar menuBar;
+  final MenuBar? menuBar;
 
   /// This will draw the menu where you tap/click on the child widget instead of at a predefined alignment.  Default is `false`
   ///
@@ -38,7 +38,7 @@ class Menu extends StatefulWidget {
   final TapType tapType;
 
   const Menu({
-    Key key,
+    Key? key,
     this.child,
     this.menuBar,
     this.menuOverTap = false,
@@ -54,12 +54,12 @@ class Menu extends StatefulWidget {
 
 class _MenuState extends State<Menu> {
   final GlobalKey key = GlobalKey();
-  OverlayEntry itemEntry;
+  OverlayEntry? itemEntry;
   final layerLink = LayerLink();
 
   bool showMenu = false;
   // ignore: prefer_final_fields
-  static List<OverlayEntry> _overlays = [];
+  static List<OverlayEntry?> _overlays = [];
   final GlobalKey sizeKey = GlobalKey();
 
   @override
@@ -72,7 +72,7 @@ class _MenuState extends State<Menu> {
       onDoubleTap: widget.tapType == TapType.doubleTap ? buildMenu : null,
       onSecondaryTapDown: (details) {
         if (widget.tapType == TapType.secondaryTap && !widget.menuOverTap) {
-          return buildMenu;
+          //return buildMenu;
         }
         return null;
       },
@@ -107,7 +107,7 @@ class _MenuState extends State<Menu> {
     );
   }
 
-  void buildMenu({Offset tapOffset}) {
+  void buildMenu({Offset? tapOffset}) {
     MenuAlignment _childAlignmentOnMenu;
     //Computes the child alignment point
     if (tapOffset != null) {
@@ -121,11 +121,13 @@ class _MenuState extends State<Menu> {
       }
     }
     // Gets the size of the parent
-    RenderBox renderObject = key.currentContext?.findRenderObject();
+    RenderBox renderObject =
+        key.currentContext?.findRenderObject() as RenderBox;
     final bounds = renderObject.paintBounds;
 
     // Gets the size of the menu
-    RenderBox menuObject = sizeKey.currentContext?.findRenderObject();
+    RenderBox menuObject =
+        sizeKey.currentContext?.findRenderObject() as RenderBox;
     final menuBounds = menuObject.paintBounds;
     Offset childOffset;
 
@@ -163,7 +165,7 @@ class _MenuState extends State<Menu> {
       ),
     );
 
-    Overlay.of(context).insert(itemEntry);
+    Overlay.of(context)!.insert(itemEntry!);
     _overlays.add(itemEntry);
   }
 
