@@ -1,7 +1,7 @@
 part of 'menu.dart';
 
 List<T> addDividers<T>(List<T> items, T divider) {
-  if (items == null) return null;
+ 
   var finalItems = <T>[];
   for (var i = 0; i < items.length; i++) {
     finalItems.add(items[i]);
@@ -29,7 +29,7 @@ class MenuBar {
 
   /// Do you want to draw a divider?
   final bool drawDivider;
-  final Color backgroundColor;
+  final Color? backgroundColor;
   final double elevation;
 
   /// This sets a border radius for the menu when [borderStyle] is set to [MenuBorderStyle.rounded]
@@ -55,10 +55,10 @@ class MenuBar {
     ],
     this.maxThickness = 36,
     this.orientation = MenuOrientation.horizontal,
-    Widget dividerWidget,
+    Widget? dividerWidget,
     this.itemPadding = const EdgeInsets.all(4),
     this.drawDivider = true,
-    Color backgroundColor,
+    Color? backgroundColor,
     this.elevation = 4,
     this.borderRadius = const BorderRadius.all(Radius.circular(16)),
     this.drawArrow = false,
@@ -76,13 +76,13 @@ class MenuBar {
 }
 
 class _MenuBar extends StatefulWidget {
-  final GlobalKey menuKey;
-  final MenuBar menuBar;
-  final MenuAlignment menuAlignment;
-  final Function dismiss;
+  final GlobalKey? menuKey;
+  final MenuBar? menuBar;
+  final MenuAlignment? menuAlignment;
+  final Function? dismiss;
 
   const _MenuBar(
-      {Key key, this.dismiss, this.menuBar, this.menuAlignment, this.menuKey})
+      {Key? key, this.dismiss, this.menuBar, this.menuAlignment, this.menuKey})
       : super(key: key);
 
   @override
@@ -90,25 +90,25 @@ class _MenuBar extends StatefulWidget {
 }
 
 class _MenuBarState extends State<_MenuBar> {
-  bool _show;
+  late bool _show;
   @override
   Widget build(BuildContext context) {
-    _show = !widget.menuBar.fadeIn;
-    final menuItems = widget.menuBar.menuItems
+    _show = !widget.menuBar!.fadeIn;
+    final menuItems = widget.menuBar!.menuItems
         .map((item) => _MenuItem(
               menuItem: item,
-              itemPadding: widget.menuBar.itemPadding,
+              itemPadding: widget.menuBar!.itemPadding,
               dismiss: widget.dismiss,
             ))
         .toList();
     // Container sets the thickness of the menu.
     return Container(
       constraints: BoxConstraints(
-          maxHeight: widget.menuBar.orientation == MenuOrientation.horizontal
-              ? widget.menuBar.maxThickness
+          maxHeight: widget.menuBar!.orientation == MenuOrientation.horizontal
+              ? widget.menuBar!.maxThickness
               : double.infinity,
-          maxWidth: widget.menuBar.orientation == MenuOrientation.vertical
-              ? widget.menuBar.maxThickness
+          maxWidth: widget.menuBar!.orientation == MenuOrientation.vertical
+              ? widget.menuBar!.maxThickness
               : double.infinity),
       alignment: Alignment.topLeft,
       child: AnimatedOpacity(
@@ -116,30 +116,30 @@ class _MenuBarState extends State<_MenuBar> {
         duration: Duration(milliseconds: 250),
         child: Material(
           key: widget.menuKey,
-          color: widget.menuBar.backgroundColor,
-          elevation: widget.menuBar.elevation,
-          borderRadius: widget.menuBar.borderStyle != MenuBorderStyle.pill
-              ? widget.menuBar.borderStyle == MenuBorderStyle.straight
-                  ? 0
-                  : widget.menuBar.borderRadius
+          color: widget.menuBar!.backgroundColor,
+          elevation: widget.menuBar!.elevation,
+          borderRadius: widget.menuBar!.borderStyle != MenuBorderStyle.pill
+              ? widget.menuBar!.borderStyle == MenuBorderStyle.straight
+                  ? 0 as BorderRadiusGeometry?
+                  : widget.menuBar!.borderRadius
               : null,
           clipBehavior: Clip.antiAlias,
-          shape: widget.menuBar.borderStyle == MenuBorderStyle.pill
+          shape: widget.menuBar!.borderStyle == MenuBorderStyle.pill
               ? MenuBorder(
                   arrowAlignment: widget.menuAlignment,
-                  drawArrow: widget.menuBar.drawArrow)
+                  drawArrow: widget.menuBar!.drawArrow)
               : null,
-          child: widget.menuBar.orientation == MenuOrientation.horizontal
+          child: widget.menuBar!.orientation == MenuOrientation.horizontal
               ? Row(
                   mainAxisSize: MainAxisSize.min,
-                  children: widget.menuBar.drawDivider
-                      ? addDividers(menuItems, widget.menuBar._divider)
+                  children: widget.menuBar!.drawDivider
+                      ? addDividers(menuItems, widget.menuBar!._divider)
                       : menuItems,
                 )
               : Column(
                   mainAxisSize: MainAxisSize.min,
-                  children: widget.menuBar.drawDivider
-                      ? addDividers(menuItems, widget.menuBar._divider)
+                  children: widget.menuBar!.drawDivider
+                      ? addDividers(menuItems, widget.menuBar!._divider)
                       : menuItems,
                 ),
         ),
@@ -150,8 +150,8 @@ class _MenuBarState extends State<_MenuBar> {
 
 class MenuBorder extends ShapeBorder {
   final bool usePadding;
-  final MenuAlignment arrowAlignment;
-  final double radius;
+  final MenuAlignment? arrowAlignment;
+  final double? radius;
   final double arrowSize;
   final bool drawArrow;
 
@@ -164,59 +164,60 @@ class MenuBorder extends ShapeBorder {
   });
 
   void trianglePathFromAlignment(
-      {Path path,
-      Rect rect,
-      double size,
-      double radius,
-      MenuAlignment alignment}) {
+      {
+        required Path path,
+      Rect? rect,
+      double? size,
+      double? radius,
+      MenuAlignment? alignment}) {
     final arrowOffset = getOffsetByAlignment(rect, alignment);
-    final cornerLength = radius ?? rect.height / 2;
+    final cornerLength = radius ?? rect!.height / 2;
     switch (alignment) {
       case MenuAlignment.topLeft:
         path
           ..moveTo(arrowOffset.dx, arrowOffset.dy + cornerLength)
-          ..lineTo(arrowOffset.dx - size, arrowOffset.dy - size)
+          ..lineTo(arrowOffset.dx - size!, arrowOffset.dy - size)
           ..lineTo(arrowOffset.dx + cornerLength, arrowOffset.dy);
         break;
       case MenuAlignment.topCenter:
         path
-          ..moveTo(arrowOffset.dx - size, arrowOffset.dy)
+          ..moveTo(arrowOffset.dx - size!, arrowOffset.dy)
           ..lineTo(arrowOffset.dx, arrowOffset.dy - size)
           ..lineTo(arrowOffset.dx + size, arrowOffset.dy);
         break;
       case MenuAlignment.topRight:
         path
           ..moveTo(arrowOffset.dx - cornerLength, arrowOffset.dy)
-          ..lineTo(arrowOffset.dx + size, arrowOffset.dy - size)
+          ..lineTo(arrowOffset.dx + size!, arrowOffset.dy - size)
           ..lineTo(arrowOffset.dx, arrowOffset.dy + cornerLength);
         break;
       case MenuAlignment.centerRight:
         path
-          ..moveTo(arrowOffset.dx, arrowOffset.dy - size)
+          ..moveTo(arrowOffset.dx, arrowOffset.dy - size!)
           ..lineTo(arrowOffset.dx + size, arrowOffset.dy)
           ..lineTo(arrowOffset.dx, arrowOffset.dy + size);
         break;
       case MenuAlignment.bottomRight:
         path
           ..moveTo(arrowOffset.dx, arrowOffset.dy - cornerLength)
-          ..lineTo(arrowOffset.dx + size, arrowOffset.dy + size)
+          ..lineTo(arrowOffset.dx + size!, arrowOffset.dy + size)
           ..lineTo(arrowOffset.dx - cornerLength, arrowOffset.dy);
         break;
       case MenuAlignment.bottomCenter:
         path
-          ..moveTo(arrowOffset.dx + size, arrowOffset.dy)
+          ..moveTo(arrowOffset.dx + size!, arrowOffset.dy)
           ..lineTo(arrowOffset.dx, arrowOffset.dy + size)
           ..lineTo(arrowOffset.dx - size, arrowOffset.dy);
         break;
       case MenuAlignment.bottomLeft:
         path
           ..moveTo(arrowOffset.dx + cornerLength, arrowOffset.dy)
-          ..lineTo(arrowOffset.dx - size, arrowOffset.dy + size)
+          ..lineTo(arrowOffset.dx - size!, arrowOffset.dy + size)
           ..lineTo(arrowOffset.dx, arrowOffset.dy - cornerLength);
         break;
       case MenuAlignment.centerLeft:
         path
-          ..moveTo(arrowOffset.dx, arrowOffset.dy + size)
+          ..moveTo(arrowOffset.dx, arrowOffset.dy + size!)
           ..lineTo(arrowOffset.dx - size, arrowOffset.dy)
           ..lineTo(arrowOffset.dx, arrowOffset.dy - size);
         break;
@@ -229,10 +230,10 @@ class MenuBorder extends ShapeBorder {
       EdgeInsets.only(bottom: usePadding ? 20 : 0);
 
   @override
-  Path getInnerPath(Rect rect, {TextDirection textDirection}) => null;
+  Path getInnerPath(Rect rect, {TextDirection? textDirection}) => Path();
 
   @override
-  Path getOuterPath(Rect rect, {TextDirection textDirection}) {
+  Path getOuterPath(Rect rect, {TextDirection? textDirection}) {
     rect = Rect.fromPoints(rect.topLeft, rect.bottomRight);
     var path = Path()
       ..addRRect(
@@ -250,7 +251,7 @@ class MenuBorder extends ShapeBorder {
   }
 
   @override
-  void paint(Canvas canvas, Rect rect, {TextDirection textDirection}) {}
+  void paint(Canvas canvas, Rect rect, {TextDirection? textDirection}) {}
 
   @override
   ShapeBorder scale(double t) => this;
